수정- 태그 이름 변경, 류오 캐릭터 생성
