+++ Config/DefaultGameplayTags.ini
@@ -12,12 +12,13 @@ InvalidTagCharacters="\"\',"
 +GameplayTagRedirects=(OldTagName="Attribute.Damage",NewTagName="Attribute.Data.Damage")
 +GameplayTagRedirects=(OldTagName="InputTag.Subskill",NewTagName="InputTag.Skill.Sub")
 +GameplayTagRedirects=(OldTagName="InputTag.MainSkill",NewTagName="InputTag.Skill.Main")
++GameplayTagRedirects=(OldTagName="Ability.Hit.Active",NewTagName="Enemy.Ability.Activating")
++GameplayTagRedirects=(OldTagName="Ability.Hit.Event",NewTagName="Enemy.Ability.TriggerAttack")
 NumBitsForContainerSize=6
 NetIndexFirstBitSegment=16
-+GameplayTagList=(Tag="Ability.Hit",DevComment="")
-+GameplayTagList=(Tag="Ability.Hit.Active",DevComment="")
-+GameplayTagList=(Tag="Ability.Hit.Event",DevComment="")
 +GameplayTagList=(Tag="Attribute.Data.Damage",DevComment="")
++GameplayTagList=(Tag="Enemy.Ability.Activating",DevComment="")
++GameplayTagList=(Tag="Enemy.Ability.TriggerAttack",DevComment="")
 +GameplayTagList=(Tag="InputTag.Action.Primary",DevComment="")
 +GameplayTagList=(Tag="InputTag.Action.Secondary",DevComment="")
 +GameplayTagList=(Tag="InputTag.Interact",DevComment="")
